backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
pypdf
python-docx
pydantic
python-dotenv
langchain
langchain-community
sentence-transformers
openai
pymupdf
